Chatbot/requirements.txt
rasa==3.6.21
rasa-sdk==3.6.2
requests==2.32.3
word2number==1.1
rapidfuzz==3.9.7

psycopg2-binary==2.9.5
SQLAlchemy==1.4.54

sentence-transformers>=2.2.2
pandas==2.0.3
numpy==1.23.5

fastapi>=0.100.0
uvicorn>=0.23.0
websockets==10.4
aiohttp==3.9.5
aiofiles==24.1.0
python-dateutil==2.8.2
ipython==7.34.0
prompt-toolkit==3.0.38
packaging==20.9
ipykernel>=6.0.0
